Disable the pie menu in custom tabs and web apps
Initialise the pie menu only if we are in the tabbed or document
activities. This also simplifies the initialisation as we no longer
search for the correct menu action method. Instead, use a bit of a
workaround to call the default onOptionsItemSelected method, which in
turn calls the corresponding method in Chrome.